+++ tui/src/ui/components/tag_editor/view.rs
@@ -2,10 +2,13 @@ use std::borrow::Cow;
 use std::path::Path;
 
 use anyhow::Result;
-use tuirealm::State;
-use tuirealm::props::{Alignment, AttrValue, Attribute, PropPayload, PropValue, Style, TextSpan};
+use tuirealm::props::{
+    AttrValue, Attribute, HorizontalAlignment, PropPayload, PropValue, SpanStatic, Style, Title,
+};
 use tuirealm::ratatui::layout::{Constraint, Layout};
 use tuirealm::ratatui::widgets::Clear;
+use tuirealm::state::State;
+use tuirealm::terminal::TerminalAdapter;
 
 use super::TETrack;
 use crate::ui::components::tag_editor::te_footer::TEFooter;
@@ -18,14 +21,14 @@ use crate::ui::model::Model;
 use crate::ui::utils::{draw_area_in_absolute, draw_area_top_right_absolute};
 
 impl Model {
+    /// Draw the Tag Editor.
     #[allow(clippy::too_many_lines)]
     pub fn view_tag_editor(&mut self) {
         self.terminal
-            .raw_mut()
             .draw(|f| {
                 let select_lyric_len =
                     match self.app.state(&Id::TagEditor(IdTagEditor::SelectLyric)) {
-                        Ok(State::One(_)) => 3,
+                        Ok(State::Single(_)) => 3,
                         _ => 8,
                     };
                 if self.app.mounted(&Id::TagEditor(IdTagEditor::LabelHint)) {
@@ -140,7 +143,7 @@ impl Model {
             .expect("Expected to draw without error");
     }
 
-    /// Mount / Remount the Tag Editor
+    /// Mount / Remount the Tag Editor.
     fn remount_tageditor(&mut self) -> Result<()> {
         self.app.remount(
             Id::Label,
@@ -187,20 +190,12 @@ impl Model {
         )?;
         self.app.remount(
             Id::TagEditor(IdTagEditor::CounterDelete),
-            Box::new(TECounterDelete::new(
-                None,
-                "Delete Selected",
-                self.config_tui.clone(),
-            )),
+            Box::new(TECounterDelete::new(None, self.config_tui.clone())),
             Vec::new(),
         )?;
         self.app.remount(
             Id::TagEditor(IdTagEditor::CounterSave),
-            Box::new(TECounterSave::new(
-                None,
-                "Save LRC",
-                self.config_tui.clone(),
-            )),
+            Box::new(TECounterSave::new(None, self.config_tui.clone())),
             Vec::new(),
         )?;
         self.app.remount(
@@ -243,7 +238,7 @@ impl Model {
         }
     }
 
-    /// Unmount the Tag Editor
+    /// Unmount the Tag Editor Components.
     fn umount_tageditor_inner(&mut self) -> Result<()> {
         self.app.umount(&Id::TagEditor(IdTagEditor::LabelHint))?;
         self.app.umount(&Id::TagEditor(IdTagEditor::InputArtist))?;
@@ -262,6 +257,7 @@ impl Model {
         Ok(())
     }
 
+    /// Unmount all tag editor components.
     pub fn umount_tageditor(&mut self) {
         self.mount_label_help();
         self.umount_tageditor_inner().unwrap();
@@ -271,12 +267,23 @@ impl Model {
     }
 
     /// Set the Lyric section of the tag-editor to the Lyrics based on the provided Track
-    #[allow(clippy::too_many_lines)]
-    pub fn init_by_song(&mut self, s: TETrack) -> Result<()> {
-        self.tageditor_song = Some(s);
-        // Unwrap safe as we literally just assigned it
-        let s = self.tageditor_song.as_ref().unwrap();
-        if let Some(artist) = s.artist() {
+    pub fn init_by_song(&mut self, track: TETrack) -> Result<()> {
+        self.tageditor.song = Some(track);
+
+        self.init_metadata()?;
+
+        self.init_lyric_data()
+    }
+
+    /// Read the metadata from the current track and set it to the Tag Editor fields.
+    ///
+    /// Expects a current `tageditor_song` to be set.
+    fn init_metadata(&mut self) -> Result<()> {
+        let Some(track) = self.tageditor.song.as_ref() else {
+            // No Metadata to change.
+            return Ok(());
+        };
+        if let Some(artist) = track.artist() {
             self.app.attr(
                 &Id::TagEditor(IdTagEditor::InputArtist),
                 Attribute::Value,
@@ -284,7 +291,7 @@ impl Model {
             )?;
         }
 
-        if let Some(title) = s.title() {
+        if let Some(title) = track.title() {
             self.app.attr(
                 &Id::TagEditor(IdTagEditor::InputTitle),
                 Attribute::Value,
@@ -292,7 +299,7 @@ impl Model {
             )?;
         }
 
-        if let Some(album) = s.album() {
+        if let Some(album) = track.album() {
             self.app.attr(
                 &Id::TagEditor(IdTagEditor::InputAlbum),
                 Attribute::Value,
@@ -300,7 +307,7 @@ impl Model {
             )?;
         }
 
-        if let Some(genre) = s.genre() {
+        if let Some(genre) = track.genre() {
             self.app.attr(
                 &Id::TagEditor(IdTagEditor::InputGenre),
                 Attribute::Value,
@@ -308,11 +315,16 @@ impl Model {
             )?;
         }
 
-        let lyric_frames = s.lyric_frames();
+        Ok(())
+    }
+
+    /// Extract the lyric data from the current track and set it in the Tag Editor fields.
+    fn init_lyric_data(&mut self) -> Result<()> {
+        let track = self.tageditor.song.as_ref().unwrap();
+        let lyric_frames = track.lyric_frames();
 
         if lyric_frames.is_empty() {
-            self.init_by_song_no_lyric();
-            return Ok(());
+            return self.init_lyric_data_empty();
         }
 
         let vec_lang: Vec<PropValue> = lyric_frames
@@ -331,14 +343,13 @@ impl Model {
             .map(PropValue::Str)
             .collect();
 
-        let selected_index = s.lyric_selected_index();
+        let selected_index = track.lyric_selected_index();
         // get access to "Lyric" instance for text and modified description for display
         let (Some(vec_lang_selected), Some(selected_desc)) =
-            (s.lyric_selected(), vec_lang.get(selected_index))
+            (track.lyric_selected(), vec_lang.get(selected_index))
         else {
             // this should not happen as if it is Some above, there should be at least one entry in the vec.
-            self.init_by_song_no_lyric();
-            return Ok(());
+            return self.init_lyric_data_empty();
         };
         let selected_desc = selected_desc
             .as_str()
@@ -349,7 +360,7 @@ impl Model {
         let vec_lyric = vec_lang_selected
             .text
             .lines()
-            .map(|line| PropValue::TextSpan(TextSpan::from(line.trim())))
+            .map(|line| PropValue::TextSpan(SpanStatic::from(line.trim().to_string())))
             .collect();
 
         self.app.attr(
@@ -360,17 +371,23 @@ impl Model {
         self.app.attr(
             &Id::TagEditor(IdTagEditor::CounterDelete),
             Attribute::Value,
-            AttrValue::Payload(PropPayload::One(PropValue::Usize(selected_index_display))),
+            AttrValue::Payload(PropPayload::Single(PropValue::Usize(
+                selected_index_display,
+            ))),
         )?;
         self.app.attr(
             &Id::TagEditor(IdTagEditor::CounterSave),
             Attribute::Value,
-            AttrValue::Payload(PropPayload::One(PropValue::Usize(selected_index_display))),
+            AttrValue::Payload(PropPayload::Single(PropValue::Usize(
+                selected_index_display,
+            ))),
         )?;
         self.app.attr(
             &Id::TagEditor(IdTagEditor::TextareaLyric),
             Attribute::Title,
-            AttrValue::Title((selected_description, Alignment::Left)),
+            AttrValue::Title(
+                Title::from(selected_description).alignment(HorizontalAlignment::Left),
+            ),
         )?;
 
         self.app.attr(
@@ -383,50 +400,31 @@ impl Model {
     }
 
     /// Set the Lyric section of the tag-editor to "No Lyrics" (ie clear state)
-    fn init_by_song_no_lyric(&mut self) {
-        assert!(
-            self.app
-                .attr(
-                    &Id::TagEditor(IdTagEditor::SelectLyric),
-                    Attribute::Content,
-                    AttrValue::Payload(PropPayload::Vec(
-                        ["Empty"]
-                            .iter()
-                            .map(|x| PropValue::Str((*x).to_string()))
-                            .collect(),
-                    )),
-                )
-                .is_ok()
-        );
-        assert!(
-            self.app
-                .attr(
-                    &Id::TagEditor(IdTagEditor::CounterDelete),
-                    Attribute::Value,
-                    AttrValue::Payload(PropPayload::None),
-                )
-                .is_ok()
-        );
+    fn init_lyric_data_empty(&mut self) -> Result<()> {
+        self.app.attr(
+            &Id::TagEditor(IdTagEditor::SelectLyric),
+            Attribute::Content,
+            AttrValue::Payload(PropPayload::Vec(vec![PropValue::Str("Empty".to_string())])),
+        )?;
+        self.app.attr(
+            &Id::TagEditor(IdTagEditor::CounterDelete),
+            Attribute::Value,
+            AttrValue::Payload(PropPayload::None),
+        )?;
 
-        assert!(
-            self.app
-                .attr(
-                    &Id::TagEditor(IdTagEditor::TextareaLyric),
-                    Attribute::Title,
-                    AttrValue::Title(("Empty Lyrics".to_string(), Alignment::Left))
-                )
-                .is_ok()
-        );
-        assert!(
-            self.app
-                .attr(
-                    &Id::TagEditor(IdTagEditor::TextareaLyric),
-                    Attribute::Text,
-                    AttrValue::Payload(PropPayload::Vec(vec![PropValue::TextSpan(
-                        TextSpan::from("No Lyrics.")
-                    ),]))
-                )
-                .is_ok()
-        );
+        self.app.attr(
+            &Id::TagEditor(IdTagEditor::TextareaLyric),
+            Attribute::Title,
+            AttrValue::Title(Title::from("Empty Lyrics").alignment(HorizontalAlignment::Left)),
+        )?;
+        self.app.attr(
+            &Id::TagEditor(IdTagEditor::TextareaLyric),
+            Attribute::Text,
+            AttrValue::Payload(PropPayload::Vec(vec![PropValue::TextSpan(
+                SpanStatic::from("No Lyrics."),
+            )])),
+        )?;
+
+        Ok(())
     }
 }